research-system/requirements.txt
# ── Core HTTP & HTML parsing ───────────────────────────────────────────────
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# ── Web search ────────────────────────────────────────────────────────────
duckduckgo-search>=6.0.0

# ── PDF processing ────────────────────────────────────────────────────────
pdfplumber>=0.10.0

# ── Templating & report ───────────────────────────────────────────────────
jinja2>=3.1.0
markdown2>=2.4.0

# ── Data ──────────────────────────────────────────────────────────────────
pandas>=2.0.0

# ── Async HTTP ────────────────────────────────────────────────────────────
aiohttp>=3.9.0

# ── CLI interface ─────────────────────────────────────────────────────────
click>=8.1.0

# ── Web app (frontend) ────────────────────────────────────────────────────
flask>=3.0.0

# ── Environment variables ─────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── LLM synthesis (Anthropic Claude) ─────────────────────────────────────
anthropic>=0.25.0
